Simplified, refactored, and cleaned up code
