+++ main.py
@@ -4,13 +4,11 @@ from collections import defaultdict
 from typeguard import check_type, TypeCheckError
 
 
+from app_interface import AppComponent
 from calculation import CalculationComponent
 from adc import ADCComponent, VirtualADCComponent
-from app_interface import AppComponent
-
 from motor import MotorComponent, VirtualMotorComponent
-
-# from nikhil_frontend import LCDController  # Please rename this package
+from pui import PUIComponent
 from ws import WebSocketComponent
 from type_defs import ADCStatus, CalculationStatus, Message, MotorStatus
 
@@ -143,14 +141,9 @@ if __name__ == "__main__":
         pub_queue=app_pub_queue, sub_queue=adc_sub_queue, motor_component=motor
     )
 
-    # === Initialize LCD Component ===
-    # lcd_sub_queue = asyncio.Queue()
-    # lcd = LCDComponent(lcd_data_queue)
-    # lcd = VirtualLCDComponent(sub_queue=lcd_sub_queue)
-
     # === Initialize Frontend ===
-    # frontend_sub_queue = asyncio.Queue()
-    # frontend = LCDController(frontend_sub_queue, app_pub_queue)
+    pui_sub_queue = asyncio.Queue()
+    pui = PUIComponent(pui_sub_queue, app_pub_queue)
 
     calculation_sub_queue = asyncio.Queue()
     calculation = CalculationComponent(
@@ -167,11 +160,12 @@ if __name__ == "__main__":
         calculation,
         motor,
         adc,
-        # frontend,
+        # pui,
     ]  # Add all components to this array
     app = App(*components, pub_queue=app_pub_queue)
 
     # === Add queue subscriptions ===
+
     app.registerSub(
         ["voltage/data", "calculation/command", "adc/status"],
         calculation_sub_queue,
@@ -186,8 +180,8 @@ if __name__ == "__main__":
     # Only uncomment this if using PiPlate or Virtual ADC
     # app.registerSub(["adc/command"], adc_sub_queue)
 
-    # Uncomment this if using Frontend
-    # app.registerSub(["signal/data"], frontend_sub_queue)
+    # Uncomment this if using physical user interface (PUI)
+    # app.registerSub(["signal/data"], pui_sub_queue)
 
     logger.info("starting app")
     asyncio.run(app.run())
