+++ motor/motor_config.py
@@ -1,7 +1,7 @@
 STEPS_PER_REV = 200
 # TODO: Define these pins
 STEP = 13  # Step pin
-DIR = 27  # Direction pin
-ENA = 22  # Enable pin
+DIR = 6  # Direction pin
+ENA = 5  # Enable pin
 
 DUTY = 50
