+++ pui/__init__.py
@@ -1,0 +1,1 @@
+from .pui_component import PUIComponent
